requirements.txt
pillow
numpy
ipython
ipdb
torchvision
tqdm
